requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.0.0
dnspython>=2.0.0